Restructure SYS_TimerTaskHandler
With these changes, the function no longer writes to halTimerIrqCount,
but stores the previous value of the count to calculate the elapsed
time. This means that disabling interupts is no longer needed.
Furthermore, this allows replacing halTimerIrqCount with a millis()
function call in the Arduino version of the hal, which is obviously
read-only.