Make inline functions inline static
This helps when compiling without optimizations turned on, in which case
actual function calls will be generated. Previously, there was no .c
file which actually provided implementations for these functions.
Instead of selecting a .c file to put them in, this just makes them
static so they end up wherever they are used.

Without optimization, this adds some overhead, but that's ok. With
optimization, these functions are almost certainly inlined due to their
small size, so there should be no additional overhead in that case.